% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advice_images.R
\name{advice_image_load}
\alias{advice_image_load}
\title{Loads one of the background images for your advice with magick package.}
\usage{
advice_image_load(image_name = "clouds")
}
\arguments{
\item{image_name}{String that specifies which background you'd like to load. The options are clouds, rainbow, rain, mountain, ocean, or paper. Default is}
}
\value{
Image object
}
\description{
Loads one of the background images for your advice with magick package.
}
