% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_advice.R
\name{load_image}
\alias{load_image}
\title{Loads one of the background images for your advice with magick package.}
\usage{
load_image(image_name = "clouds")
}
\arguments{
\item{image_name}{String that specifies which background you'd like to load. The options are clouds, rainbow, mountain, ocean, sunset, or paper. Default is}
}
\value{
Image object
}
\description{
Loads one of the background images for your advice with magick package.
}
\examples{
load_image(image_name = "clouds")

}
