% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advice_images.R
\name{print_advice}
\alias{print_advice}
\title{Prints advice on your background image (from advicegiveR::image_load())}
\usage{
print_advice(image, advice, textcolor = "white", size = 30)
}
\arguments{
\item{image}{Image object}

\item{advice}{Advice string from advicegiveR::get_advice() or a self-supplied string}

\item{textcolor}{The color of the advice. Accepts color strings or hex codes. Default is white.}

\item{size}{Size of the text. Default is 30.}
}
\value{
Image with the advice message annotated on top of it
}
\description{
Prints advice on your background image (from advicegiveR::image_load())
}
\examples{
print_advice(image = background, advice = advice_text, textcolor = "yellow", size = 35)
}
