% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_advice.R
\name{get_advice}
\alias{get_advice}
\title{This function makes a call to the advice slip API to return a random piece of advice. Requires jsonlite.}
\usage{
get_advice(id = sample(0:217, 1))
}
\arguments{
\item{id}{An integer between 1 and 216. Default is randomly generated.}
}
\value{
A string of advice
}
\description{
This function makes a call to the advice slip API to return a random piece of advice. Requires jsonlite.
}
